+++ nion/ui/QtUserInterface.py
@@ -740,6 +740,15 @@ class QtStackWidgetBehavior(QtWidgetBehavior):
     def current_index(self, index: int) -> None:
         self.__current_index = index
         self.proxy.StackWidget_setCurrentIndex(self.widget, index)
+        # see sizing notes:
+        # https://wiki.qt.io/Technical_FAQ#How_can_I_get_a_QStackedWidget_to_automatically_switch_size_depending_on_the_content_of_the_page.3F
+        # https://stackoverflow.com/questions/14480696/resize-qstackedwidget-to-the-page-which-is-opened
+        for i in range(self.proxy.Widget_widgetCount(self.widget)):
+            widget = self.proxy.Widget_widgetByIndex(self.widget, i)
+            self.proxy.Widget_setWidgetProperty(widget, "size-policy-horizontal", "preferred" if i == index else "ignored")
+            self.proxy.Widget_setWidgetProperty(widget, "size-policy-vertical", "preferred" if i == index else "ignored")
+            self.proxy.Widget_adjustSize(widget)
+        self.proxy.Widget_adjustSize(self.widget)
 
 
 class QtGroupWidgetBehavior(QtWidgetBehavior):